+++ CHANGES.md
@@ -48,6 +48,7 @@ to extends ImageryProvider, also add a corresponding object to all that extends 
 * changed class DynamicGeometryUpdater to trait DynamicGeometryUpdater
 * added rotatable2D to Viewer, Scene, CesiumWidget, ViewerOptions, SceneOptions, CesiumWidgetOptions
 * added packArray and unpackArray to Cartesian2, Cartesian3, and Cartesian4 objects
-
+* added aircraftHeadingPitchRollToFixedFrame, aircraftHeadingPitchRollQuaternion to Transforms
+* renamed headingPitchRollMatrix4 to headingPitchRollToFixedFrame in Transforms